requirements.txt
https://github.com/datahq/planner/archive/master.zip
https://github.com/datahq/specstore/archive/master.zip
https://github.com/datahq/events/archive/master.zip
https://github.com/datahq/filemanager/archive/master.zip
datapackage-pipelines-elasticsearch>=0.0.6
datapackage-pipelines-aws>=0.0.19
goodtables>=1.5.1
